Add files via upload
this is my first task of my internship.
i create this project using the python programming.
In this project the tkinter library is used. Tkinter is used for making GUI.
Also the if and else statements are used for list of task.
The end user after the running the code he/she simply add the taskand also delete the task.